+++ R/extract_geocoverage.R
@@ -53,48 +53,6 @@
 #'     An existing geographic coverage template will not be overwritten by 
 #'     subsequent calls to \code{extract_geocoverage()}.
 #'     
-#' @examples 
-#' # Set working directory
-#' setwd(tempdir())
-#' 
-#' # Create data package directory "pkg_255"
-#' file.copy(
-#'   from = system.file('/examples/pkg_255', package = 'EMLassemblyline'),
-#'   to = '.',
-#'   recursive = TRUE
-#' )
-#' 
-#' # View directory contents (NOTE: geographic_coverage.txt doesn't exist)
-#' dir('./pkg_255/metadata_templates')
-#' 
-#' # Extract geographic coverage
-#' extract_geocoverage(
-#'   path = './pkg_255/metadata_templates',
-#'   data.path = './pkg_255/data_objects',
-#'   data.table = 'nitrogen.csv',
-#'   site.col = 'site_name',
-#'   lat.col = 'site_lat',
-#'   lon.col = 'site_lon'
-#' )
-#' 
-#' # View directory contents (NOTE: geographic_coverage.txt exists)
-#' dir('./pkg_255/metadata_templates')
-#' 
-#' # Rerunning extract_geocoverage() does not overwrite files
-#' extract_geocoverage(
-#'   path = './pkg_255/metadata_templates',
-#'   data.path = './pkg_255/data_objects',
-#'   data.table = 'nitrogen.csv',
-#'   site.col = 'site_name',
-#'   lat.col = 'site_lat',
-#'   lon.col = 'site_lon'
-#' )
-#' 
-#' # Clean up
-#' unlink(
-#'   './pkg_255',
-#'   recursive = TRUE
-#' )
 #'     
 #' @export
 #'