+++ R/make_eml.R
@@ -55,7 +55,8 @@
 #'     are negative. Don't use this argument if geographic coverage is supplied
 #'     by geographic_coverage.txt.
 #' @param maintenance.description
-#'     (character) Data collection status ("ongoing" or "complete").
+#'     (character) A description of data collection status (e.g. "ongoing", 
+#'     "complete"), communicating the frequency of updates.
 #' @param data.table
 #'     (character; optional) Table file name. If more than one, then supply 
 #'     as a vector of character strings (e.g. 
@@ -78,7 +79,9 @@
 #'     (character; optional) The publicly accessible URL from which 
 #'     \code{data.table} can be downloaded. If more than one, then supply as 
 #'     a vector of character strings in the same order as listed in 
-#'     \code{data.table}.
+#'     \code{data.table}. If wanting to include URLs for some but not all
+#'     \code{data.table}, then use a "" for those that don't have a URL
+#'     (e.g. \code{data.table.url = c("", "/url/to/decomp.csv")}).
 #' @param other.entity
 #'     (character; optional) Name of \code{other.entity}(s) in this 
 #'     dataset. Use \code{other.entity} for all non-\code{data.table} files. 
@@ -98,21 +101,36 @@
 #'     (character; optional) The publicly accessible URL from which 
 #'     \code{other.entity} can be downloaded. If more than one, then supply as 
 #'     a vector of character strings in the same order as listed in 
-#'     \code{other.entity}.
+#'     \code{other.entity}. If wanting to include URLs for some but not all
+#'     \code{other.entity}, then use a "" for those that don't have a URL
+#'     (e.g. \code{other.entity.url = c("", "/url/to/quality_control.R")}).
 #' @param provenance
 #'     (character; optional) EDI Data Repository Data package ID(s) 
 #'     corresponding to parent datasets from which this dataset was created 
 #'     (e.g. \code{knb-lter-cap.46.3}).
 #' @param user.id
-#'     (character; optional) ID(s) of data repository user account(s). If more 
-#'     than one, supply as a vector of character strings.
+#'     (character; optional) Repository user identifier. If more than one, 
+#'     then enter as a vector of character strings (e.g. 
+#'     \code{c("user_id_1", "user_id_2")}). \code{user.id} sets the 
+#'     /eml/access/principal element for all \code{user.domain} except
+#'     "KNB", "ADC", and if \code{user.domain = NULL}.
 #' @param user.domain
-#'     (character; optional) Domain of the \code{user.id}(s). Valid options 
-#'     for EDI are "LTER" and "EDI". If more than one, supply as a vector of 
-#'     character strings in the same order as corresponding \code{user.id}(s).
+#'     (character; optional) Repository domain associated with 
+#'     \code{user.id}. Currently supported values are "EDI" 
+#'     (Environmental Data Initiative), "LTER" (Long-Term Ecological Research 
+#'     Network), "KNB" (The Knowledge Network for Biocomplexity), "ADC" (The 
+#'     Arctic Data Center). If you'd like your system supported please contact
+#'     maintainers of the EMLassemblyline R package. If using more than one 
+#'     \code{user.domain}, then enter as a vector of character strings (e.g. 
+#'     \code{c("user_domain_1", "user_domain_2")}) in the same order as 
+#'     corresponding \code{user.id}. If \code{user.domain} is missing then a 
+#'     default value "unknown" is assigned. \code{user.domain} sets the EML 
+#'     header "system" attribute and for all \code{user.domain}, except "KNB" 
+#'     and "ADC", sets the /eml/access/principal element attributes and values.
 #' @param package.id
-#'     (character; optional) Data Repository data package ID for this dataset. A 
-#'     missing package ID defaults to "edi.101.1".
+#'     (character; optional) Data package ID for the dataset described by this
+#'     EML. Ask your data repository for a package ID. Missing \code{package.id}
+#'     is assigned a UUID.
 #' @param write.file
 #'     (logical; optional) Whether to write the EML file.
 #' @param return.obj
@@ -181,7 +199,7 @@ make_eml <- function(
   temporal.coverage,
   geographic.description, 
   geographic.coordinates, 
-  maintenance.description, 
+  maintenance.description = NULL, 
   data.table = NULL, 
   data.table.name = data.table,
   data.table.description = NULL, 
@@ -279,7 +297,7 @@ make_eml <- function(
     data.url <- data.files.url
   }
   # FIXME: Do not remove until March 2021
-  if (!missing(data.url)){
+  if (!is.null(data.url)){
     warning(
       paste0("Argument 'data.url' is deprecated; please use 'data.table.url' ",
              "and 'other.entity.url' instead."),
@@ -344,6 +362,7 @@ make_eml <- function(
   # attributes.txt templates are input as files. In contrast, when inputs are 
   # supplied by the argument "x" and when is.na() returns TRUE for values in
   # the missingValueCode field, these NAs are converted to "".
+  # TODO: Move this process to validate_templates()
   
   for (k in names(x$template)) {
     if (is.data.frame(x$template[[k]]$content)) {
@@ -364,7 +383,7 @@ make_eml <- function(
   # Validate templates --------------------------------------------------------
   
   x <- remove_empty_templates(x)
-  validate_templates("make_eml", x)
+  x <- validate_templates("make_eml", x)
 
   # Modify templates ----------------------------------------------------------
   # Modification of some template content helps with downstream processes.
@@ -593,7 +612,38 @@ make_eml <- function(
         how = c("replace"))
       rp_personnel
       
-    } else {
+    } else if (person_role == "publisher") {
+      
+      rp_personnel <- list(
+        individualName = list(
+          givenName = list(
+            x$template$personnel.txt$content[info_row,"givenName"],
+            x$template$personnel.txt$content[info_row,"middleInitial"]),
+          surName = x$template$personnel.txt$content[info_row,"surName"]),
+        organizationName = x$template$personnel.txt$content[info_row,"organizationName"],
+        electronicMailAddress = x$template$personnel.txt$content[info_row,"electronicMailAddress"])
+      if (nchar(x$template$personnel.txt$content[info_row,"userId"]) == 19){
+        rp_personnel$userId <- list(
+          directory = 'https://orcid.org',
+          paste0(
+            "https://orcid.org/", 
+            x$template$personnel.txt$content[info_row,"userId"]))
+      }
+      # FIXME Blank entries ('') result in closing tags when EML is written
+      # to file. Need function to set all elements of value = '' to NULL.
+      rp_personnel <- rapply(
+        rp_personnel,
+        function(x){
+          if (x == ""){
+            x <- NULL
+          } else {
+            x
+          }
+        },
+        how = c("replace"))
+      rp_personnel
+      
+    }else {
       
       # If givenName, middleName, and surName are blank then the 
       # associatedParty is an organization, otherwise the associatedParty 
@@ -655,68 +705,98 @@ make_eml <- function(
   
   # Create <eml> --------------------------------------------------------------
   
-  # FIXME: Support other system values
-  
   message("Making EML ...")
   message("<eml>")
   
-  if (is.null(package.id)){
-    package.id <- 'edi.101.1'
+  # Set default package.id
+  
+  if (is.null(package.id)) {
+    warning("Missing 'package.id', assigning a UUID", call. = FALSE)
+    package.id <- uuid::UUIDgenerate()
   }
+  
+  # Define the EML system attribute from user.domain
+  
+  if (("edi" %in% tolower(user.domain)) | ("lter" %in% tolower(user.domain))) {
+    sys <- "edi"
+  } else if ("knb" %in% tolower(user.domain)) {
+    sys <- "knb"
+  } else if ("adc" %in% tolower(user.domain)) {
+    sys <- "https://arcticdata.io"
+  } else {
+    sys <- "unknown"
+  }
+  
+  # Construct EML header
   
   eml <- list(
     schemaLocation = "https://eml.ecoinformatics.org/eml-2.2.0  https://nis.lternet.edu/schemas/EML/eml-2.2.0/xsd/eml.xsd",
     packageId = package.id,
-    system = "edi")
+    system = sys)
   
   # Create <access> -----------------------------------------------------------
+  # Create the access node for systems/repositories (i.e. user.domain) that 
+  # use it, and if a user.id is specified.
   
-  # FIXME: Support other user.id, user.domain, authSystem,
-  
-  message("  <access>")
-  
-  # Set default user.id and user.domain
-  
-  if (is.null(user.id)) {
-    warning(
-      "No 'user.id' was supplied. The default 'someuserid' will be used.",
-      call. = F)
-    user.id <- "someuserid"
-  } else if (is.null(user.domain)) {
-    warning(
-      "No 'user.domain' was supplied. The default 'user.domain' will be used.",
-      call. = F)
-    user.domain <- "someuserdomain"
-  }
-  
-  # Initialize the <access> node
-  
-  eml$access <- list(
-    scope = "document",
-    order = "allowFirst",
-    authSystem = "https://pasta.edirepository.org/authentication",
-    allow = list())
-  
-  # Set permissions to "all" for the EML creator/manager and "read" for 
-  # everyone else
-  
-  r <- lapply(
-    seq_along(user.id),
-    function(k) {
-      if (user.domain[k] == "LTER") {
-        principal <- paste0(
-          "uid=", user.id[k], ",o=", user.domain[k], ",dc=ecoinformatics,dc=org")
-      } else if (user.domain[k] == "EDI") {
-        principal <- paste0(
-          "uid=", user.id[k], ",o=", user.domain[k], ",dc=edirepository,dc=org")
-      } else {
-        principal <- user.id
-      }
-      list(principal = principal, permission = "all")
+  if ((eml$system != "knb") & 
+      (eml$system != "https://arcticdata.io") &
+      (!is.null(user.id))) {
+    
+    message("  <access>")
+    
+    # Set access attributes
+    
+    if (eml$system == "edi") {
+      auth_sys <- "https://pasta.edirepository.org/authentication"
+    } else {
+      auth_sys <- "unknown"
     }
-  )
-  r[[length(r)+1]] <- list(principal = "public", permission = "read")
-  eml$access$allow <- r
+    
+    eml$access <- list(
+      scope = "document",
+      order = "allowFirst",
+      authSystem = auth_sys,
+      allow = list())
+    
+    # Extrapolate user.id and user.domain if unequal lengths.
+    
+    if (!is.null(user.id) & !is.null(user.domain)) {
+      if (length(user.id) != length(user.domain)) {
+        user.id <- data.frame(
+          user.id, user.domain, stringsAsFactors = FALSE)$user.id
+        user.domain <- data.frame(
+          user.id, user.domain, stringsAsFactors = FALSE)$user.domain
+        warning(
+          "The number of 'user.id' and 'user.domain' inputs differ. ",
+          "Extrapolating to make equal:\nuser.id = ", 
+          paste(user.id, collapse = ", "), "\nuser.domain = ", 
+          paste(user.domain, collapse = ", "), call. = FALSE)
+      }
+    }
+    
+    # Set permissions to "all" for the principal and "read" for public
+    
+    r <- lapply(
+      seq_along(user.id),
+      function(k) {
+        if (auth_sys == "https://pasta.edirepository.org/authentication") {
+          if (tolower(user.domain[k]) == "lter") {
+            principal <- paste0(
+              "uid=", user.id[k], ",o=", user.domain[k], ",dc=ecoinformatics,dc=org")
+          } else if (tolower(user.domain[k]) == "edi") {
+            principal <- paste0(
+              "uid=", user.id[k], ",o=", user.domain[k], ",dc=edirepository,dc=org")
+          }
+        } else {
+          principal <- user.id[k]
+        }
+        permission <- "all"
+        list(principal = principal, permission = permission)
+      })
+    r[[length(r)+1]] <- list(principal = "public", permission = "read")
+    eml$access$allow <- r
+    
+  }
 
   # Create <dataset> ----------------------------------------------------------
   # Initialize the dataset list to which sub-nodes will be added
@@ -944,7 +1024,9 @@ make_eml <- function(
   # Create <maintenance> ------------------------------------------------------
   
   message("    <maintenance>")
-  eml$dataset$maintenance$description <- maintenance.description
+  if (!is.null(maintenance.description)) {
+    eml$dataset$maintenance$description <- maintenance.description
+  }
 
   # Create <contact> ----------------------------------------------------------
 
@@ -953,6 +1035,15 @@ make_eml <- function(
     function(k) {
       message("    <contact>")
       set_person(info_row = k, person_role = "contact")
+    })
+  
+  # Create <publisher> --------------------------------------------------------
+  
+  eml$dataset$publisher <- lapply(
+    which(x$template$personnel.txt$content$role == "publisher"),
+    function(k) {
+      message("    <publisher>")
+      set_person(info_row = k, person_role = "publisher")
     })
 
   # Create <methods> ----------------------------------------------------------
@@ -1179,8 +1270,15 @@ make_eml <- function(
         if (!is.null(data.url)) {
           physical$distribution$online$url[[1]] <- paste0(data.url, "/", k)
         } else if (!is.null(data.table.url)) {
-          physical$distribution$online$url <- data.table.url[
-            which(k == names(x$data.table))]
+          # data.table.url isn't required for each data table, but must have a 
+          # non-NULL entry in the data.table.url if other data tables have a 
+          # URL. A "" or NA indicates to skip URL assignment.
+          url <- data.table.url[which(k == names(x$data.table))]
+          if ((url == "") | is.na(url)) {
+            physical$distribution <- list()
+          } else {
+            physical$distribution$online$url <- url
+          }
         } else {
           physical$distribution <- list()
         }
@@ -1217,7 +1315,6 @@ make_eml <- function(
   }
   
   # Create <otherEntity> ------------------------------------------------------
-  # FIXME: Need methods for inferring entity type, format, and other metadata
   
   if (!is.null(x$other.entity)) {
     eml$dataset$otherEntity <- lapply(
@@ -1229,14 +1326,22 @@ make_eml <- function(
           EML::set_physical(
             paste0(data.path, "/", k)))
         physical$dataFormat$textFormat <- NULL
-        physical$dataFormat$externallyDefinedFormat$formatName <- "unknown"
+        physical$dataFormat$externallyDefinedFormat$formatName <- mime::guess_type(
+          file = k, unknown = "Unknown", empty = "Unknown")
         # FIXME: data.url is deprecated. Remove support for this argument after (11 March 2021)
         if (!is.null(data.url)) {
           physical$distribution$online$url[[1]] <- paste0(data.url, "/", k)
         } else if (!is.null(other.entity.url)) {
-          physical$distribution$online$url <- other.entity.url[
-            which(k == names(x$other.entity))]
-        }else {
+          # other.entity.url isn't required for each data table, but must have a 
+          # non-NULL entry in the other.entity.url if other data tables have a 
+          # URL. A "" or NA indicates to skip URL assignment.
+          url <- other.entity.url[which(k == names(x$other.entity))]
+          if ((url == "") | is.na(url)) {
+            physical$distribution <- list()
+          } else {
+            physical$distribution$online$url <- url
+          }
+        } else {
           physical$distribution <- list()
         }
         # Create otherEntity