+++ NEWS.md
@@ -1,3 +1,33 @@
+# EMLassemblyline 2.17.0
+
+### Enhancements
+
+* __Create EML for non-EDI repositories:__ Create EML for non-EDI repositories by refactoring the logic around the `make_eml()` function arguments `user.id`, `user.domain`, and `package.id`. Details are listed in the function documentation.
+
+# EMLassemblyline 2.16.1
+
+### Bug fix
+
+* __template reader:__ Fixed a bug in the metadata template reader.
+
+# EMLassemblyline 2.16.0
+
+### Enhancements
+
+* __maintenance.description:__ The `maintenance.description` of `make_eml()` is no longer required however, a missing `maintenance.description` will return a warning with the recommended best practice.
+* __publisher:__ A data publisher can now be added by listing the person (or organization) with a "publisher" role to the personnel.txt template.
+* __project:__ Missing project information (i.e. Principal Investigator and project metadata) return a warning with the recommended practice.
+* __formatName:__ The formatName of an otherEntity is now auto-detected using the `mime` library. Undetected MIME Types are listed as "Unknown". Fixes [issue #68](https://github.com/EDIorg/EMLassemblyline/issues/68).
+* __distribution:__ Previously, when assigning a .//physical/distribtuion/online/url for two or more data tables or other entities, each was required to have a corresponding URL listed under the `make_eml()` arguments `data.table.url` and `other.entity.url` . Some use cases require assignment of a URL to only one in a list of two or more. This constraint as been relaxed so if a data object doesn't have a corresponding URL then use the values `""` or `NA` (e.g. if in `make_eml()` the argument `data.table = c("nitrogen.csv", "decomp.csv")`, and a URL only exists for the second object, then `data.table.url = c("", "/url/to/decomp.csv")`.
+
+# EMLassemblyline 2.15.0
+
+### Enhancements
+
+* __Installation:__ Simplified instructions so dependencies will be installed but and users will not be asked to upgrade installed packages (a point of confusion among many).
+
+* __Default false numeric attributes to character:__ Default user specified numeric attributes to character class when the attribute contains character values outside of that specified under the missingValueCode field of the attributes.txt template. A warning alerts the user of the issue and preserves the original data by not coercing to numeric.
+
 # EMLassemblyline 2.14.0
 
 ### Enhancement